move org stats card to discrete file
